+++ History.md
@@ -2,6 +2,7 @@
 
 * Update chai (1.1.1)
 * Improved error messaging (rake konacha:run)
+* Show pending specs in `konacha:run` output
 
 # 1.4.1